# SelfLearning/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blog Q&A API Service",
    "description": "\n    API for blog content processing, question generation, and retrieval.\n    \n    ## Authentication\n    \n    This API uses two types of API key authentication:\n    \n    ### Admin Authentication (X-Admin-Key)\n    - Required for all `/api/v1/publishers/*` endpoints\n    - Used for publisher management and administration\n    - Header: `X-Admin-Key: admin_your-key-here`\n    \n    ### Publisher Authentication (X-API-Key)\n    - Required for content endpoints: `/api/v1/questions/by-url`, `/api/v1/jobs/process`\n    - Each publisher gets a unique API key upon onboarding\n    - Validates domain ownership automatically\n    - Header: `X-API-Key: pub_your-key-here`\n    \n    ## Features\n    - Automated blog content processing\n    - AI-generated questions and summaries\n    - Vector similarity search\n    - Custom Q&A endpoint\n    - Publisher management and rate limiting\n    ",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/questions/by-url": {
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "Get Questions By Url",
        "description": "Get all questions for a specific blog URL.\n\n**Authentication**: Requires X-API-Key header with valid publisher API key.\n\nQuestions are randomized on each request for better user experience.\n\nReturns standardized format:\n{\n    \"status\": \"success\",\n    \"status_code\": 200,\n    \"message\": \"Questions retrieved successfully\",\n    \"result\": {...},\n    \"request_id\": \"req_abc123\",\n    \"timestamp\": \"2025-10-18T14:30:00.123Z\"\n}",
        "operationId": "get_questions_by_url_api_v1_questions_by_url_get",
        "parameters": [
          {
            "name": "blog_url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The blog URL to get questions for",
              "title": "Blog Url"
            },
            "description": "The blog URL to get questions for"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Publisher API key for accessing content",
              "title": "X-Api-Key"
            },
            "description": "Publisher API key for accessing content"
          }
        ],
        "responses": {
          "200": {
            "description": "Questions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionsByUrlResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required - X-API-Key header missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Publisher account not active or domain mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No questions found for the given URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/questions/{question_id}": {
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "Get Question By Id",
        "description": "Get a specific question by ID.\n\n**Authentication**: Requires X-API-Key header with valid publisher API key.\n\nReturns standardized format:\n{\n    \"status\": \"success\",\n    \"status_code\": 200,\n    \"message\": \"Question retrieved successfully\",\n    \"result\": {...},\n    \"request_id\": \"req_abc123\",\n    \"timestamp\": \"2025-10-18T14:30:00.123Z\"\n}",
        "operationId": "get_question_by_id_api_v1_questions__question_id__get",
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Publisher API key for accessing content",
              "title": "X-Api-Key"
            },
            "description": "Publisher API key for accessing content"
          }
        ],
        "responses": {
          "200": {
            "description": "Question retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionByIdResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required - X-API-Key header missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Publisher account not active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Question not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BlogInfoSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "example": "68f216e74c1c51f257077316"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "example": "Article Title"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "example": "https://example.com/article"
          },
          "author": {
            "type": "string",
            "title": "Author",
            "example": "John Doe"
          },
          "published_date": {
            "type": "string",
            "title": "Published Date",
            "example": "2025-10-18"
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count",
            "example": 10
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "url",
          "author",
          "published_date",
          "question_count"
        ],
        "title": "BlogInfoSchema",
        "description": "Blog info schema."
      },
...
      "ErrorDetailSchema": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "example": "NOT_FOUND"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "example": "Resource not found"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field",
            "example": "blog_url"
          }
        },
        "type": "object",
        "required": [
          "code",
          "detail"
        ],
        "title": "ErrorDetailSchema",
        "description": "Error detail schema for Swagger."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuestionByIdResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "success",
            "example": "success"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code",
            "example": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "example": "Operation successful"
          },
          "result": {
            "$ref": "#/components/schemas/QuestionSchema"
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseMetadataSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "warnings": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warnings"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "example": "req_abc123def456"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "example": "2025-10-18T14:30:00.123456"
          }
        },
        "type": "object",
        "required": [
          "status_code",
          "message",
          "result",
          "request_id",
          "timestamp"
        ],
        "title": "QuestionByIdResponse",
        "description": "Response for GET /questions/{question_id}."
      },
      "QuestionSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "example": "68f216e74c1c51f257077318"
          },
          "question": {
            "type": "string",
            "title": "Question",
            "example": "What is the main topic of this article?"
          },
          "answer": {
            "type": "string",
            "title": "Answer",
            "example": "The article discusses..."
          },
          "icon": {
            "type": "string",
            "title": "Icon",
            "example": "💡"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "example": "2025-10-18T14:30:00"
          }
        },
        "type": "object",
        "required": [
          "id",
          "question",
          "answer",
          "icon",
          "created_at"
        ],
        "title": "QuestionSchema",
        "description": "Question schema."
      },
      "QuestionsByUrlResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "success",
            "example": "success"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code",
            "example": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "example": "Operation successful"
          },
          "result": {
            "$ref": "#/components/schemas/QuestionsByUrlResult"
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponseMetadataSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "warnings": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warnings"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "example": "req_abc123def456"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "example": "2025-10-18T14:30:00.123456"
          }
        },
        "type": "object",
        "required": [
          "status_code",
          "message",
          "result",
          "request_id",
          "timestamp"
        ],
        "title": "QuestionsByUrlResponse",
        "description": "Response for GET /questions/by-url."
      },
      "QuestionsByUrlResult": {
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionSchema"
            },
            "type": "array",
            "title": "Questions"
          },
          "blog_info": {
            "$ref": "#/components/schemas/BlogInfoSchema"
          }
        },
        "type": "object",
        "required": [
          "questions",
          "blog_info"
        ],
        "title": "QuestionsByUrlResult",
        "description": "Result for questions by URL endpoint."
      },
...
      "ResponseMetadataSchema": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "example": 100
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "example": 1
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size",
            "example": 10
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "page_size"
        ],
        "title": "ResponseMetadataSchema",
        "description": "Metadata schema for list responses."
      },
...
      "StandardErrorResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "error",
            "example": "error"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code",
            "example": 404
          },
          "message": {
            "type": "string",
            "title": "Message",
            "example": "Resource not found"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetailSchema"
          },
          "warnings": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warnings"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "example": "req_abc123def456"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "example": "2025-10-18T14:30:00.123456"
          }
        },
        "type": "object",
        "required": [
          "status_code",
          "message",
          "error",
          "request_id",
          "timestamp"
        ],
        "title": "StandardErrorResponse",
        "description": "Standard error response schema."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "AdminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Key",
        "description": "Admin API key for publisher management endpoints"
      },
      "PublisherKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Publisher API key for content access and processing"
      }
    }
  }
}